+++ desktop/cmp/grid/impl/filter/headerfilter/HeaderFilterModel.ts
@@ -7,10 +7,20 @@
 
 import {TabContainerModel} from '@xh/hoist/cmp/tab';
 import {HoistModel, managed, lookup} from '@xh/hoist/core';
+import {
+    CompoundFilter,
+    FieldFilter,
+    FieldType,
+    Filter,
+    FilterLike,
+    flattenFilter,
+    parseFilter,
+    Store
+} from '@xh/hoist/data';
 import {action, computed} from '@xh/hoist/mobx';
 import {wait} from '@xh/hoist/promise';
-import {isEmpty} from 'lodash';
-import {GridFilterFieldSpec} from '@xh/hoist/cmp/grid';
+import {isEmpty, isEqual} from 'lodash';
+import {GridFilterFieldSpec, GridFilterModel} from '@xh/hoist/cmp/grid';
 import {customTab} from './custom/CustomTab';
 import {CustomTabModel} from './custom/CustomTabModel';
 import {valuesTab} from './values/ValuesTab';
@@ -29,46 +39,52 @@ export class HeaderFilterModel extends HoistModel {
     @managed valuesTabModel: ValuesTabModel;
     @managed customTabModel: CustomTabModel;
 
-    // Set to `false` when `esc` is pressed to close without committing.
-    shouldCommitOnClose = true;
-
-    get filterModel() {
+    get filterModel(): GridFilterModel {
         return this.parent.filterModel;
     }
 
-    get field() {
+    get field(): string {
         return this.fieldSpec.field;
     }
 
-    get store() {
+    get store(): Store {
         return this.filterModel.gridModel.store;
     }
 
-    get fieldType() {
+    get fieldType(): FieldType {
         return this.store.getField(this.field).type;
     }
 
-    get currentGridFilter() {
+    get currentGridFilter(): Filter {
         return this.filterModel.filter;
     }
 
-    get columnFilters() {
+    get columnFilters(): FieldFilter[] {
         return this.filterModel.getColumnFilters(this.field);
     }
 
-    get columnCompoundFilter() {
+    get columnCompoundFilter(): CompoundFilter {
         return this.filterModel.getColumnCompoundFilter(this.field);
     }
 
-    get hasFilter() {
+    get hasFilter(): boolean {
         return !isEmpty(this.columnFilters);
     }
 
-    get hasPendingFilter() {
+    get hasPendingFilter(): boolean {
+        return !!this.pendingFilter;
+    }
+
+    @computed
+    get isDirty(): boolean {
+        return !isEqual(this.columnFilters, flattenFilter(parseFilter(this.pendingFilter)));
+    }
+
+    get pendingFilter(): FilterLike {
         const {activeTabId} = this.tabContainerModel;
         return activeTabId === 'valuesFilter'
-            ? !!this.valuesTabModel.filter
-            : !!this.customTabModel.filter;
+            ? this.valuesTabModel.filter
+            : this.customTabModel.filter;
     }
 
     @computed
@@ -121,13 +137,6 @@ export class HeaderFilterModel extends HoistModel {
         });
 
         this.syncWithFilter();
-    }
-
-    override destroy() {
-        super.destroy();
-        if (this.shouldCommitOnClose && (this.hasFilter || this.hasPendingFilter)) {
-            this.commit(false);
-        }
     }
 
     @action